backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.76
scipy==1.11.1
Pillow==10.0.0
matplotlib==3.7.1
scikit-learn==1.3.0
tqdm==4.65.0
h5py==3.9.0